\name{BigTSP-package}
\alias{BigTSP-package}
\alias{BigTSP}
\docType{package}
\title{
Top Scoring Pair based methods for classification.
}
\description{
This package is trying to implement Top Scoring Pair based methods for classification including LDCA, TSP-tree, TSP-random forest and TSP gradient boosting algorithm.
}
\details{
\tabular{ll}{
Package: \tab BigTSP\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-08-20\cr
License: \tab GPL(>= 2)\cr
}
LDCA, tsp.tree, tsp.randomForest, tsp.gbm
}
\author{
Xiaolin Yang, Han Liu

Maintainer: Who to complain to <xyang@stat.cmu.edu>
Xiaolin Yang
}

\keyword{ package }
